Don't keep invalid .caps
We do not want empty captures or such without WPA2 signatures either.

+++ src/airba.sh
@@ -14,11 +14,13 @@ airocheck() {
   for s in $(ls "$1"*.cap 2>/dev/null); do
     if [ $("$AIRCRACK_BIN" -a 2 "$s" -w "$path$wl"known.txt 2>/dev/null | grep -c valid) -eq 0 ]; then
       airstrtmp=$("$AIRCRACK_BIN" -a 2 "$s" -w "$path$wl"known.txt 2>/dev/null | grep handshake)
-      numhs=$(echo "$airstrtmp" | awk '{print $5}')
-      if [ "$numhs" == "(0" ]; then
+      if [ "$airstrtmp" == "" ]; then
         cs=0
-      else
-        cs=1
+      else      
+        numhstmp=$(echo "$airstrtmp" | awk '{print $5}')
+        if [ "$numhstmp" != "(0" ]; then
+         cs=1
+        fi
       fi
     fi
   done
@@ -156,6 +158,9 @@ ctr3=0
 for h in $(ls "$path"*.cap 2>/dev/null); do
   if [ $("$AIRCRACK_BIN" -a 2 "$h" -w "$path$wl"known.txt 2>/dev/null | grep -c valid) -eq 0 ]; then
     airstr=$("$AIRCRACK_BIN" -a 2 "$h" -w "$path$wl"known.txt 2>/dev/null | grep handshake)
+    if [ "$airstr" == "" ]; then
+      continue
+    fi
     getloc=$(sed '2q;d' "$path$loc"*.txt 2>/dev/null)
     if [ $(echo "$airstr" | awk '{print $5}') == "(0" ]; then
       rm -f "$h"
